Update ascii-mirror-bracket implementation
Test for char? instead of integer? which is consistent with how the
other procedures are implemented. This has implications when auto-
converting code to use fixnum math: (not (char? x)) matches all
integers (including bignums) whereas (fixnum? x) does not recognize
bignums. We want to recognize bignums and give them to the fx* fixnum
procedures which will then raise an error.